Moving classes to crud layer.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3934 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/X12/EB271.cs
@@ -125,7 +125,7 @@ namespace OpenDentBusiness
 			}
 			//10-Quantity
 			if(Segment.Get(10)!="") {
-				Benefitt.Quantity=PIn.Int(Segment.Get(10));
+				Benefitt.Quantity=PIn.Byte(Segment.Get(10));
 			}
 			//11-Authorization. Ignored.
 			//12-In network. Ignored.
